ngr_website/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cleanup==5.2.0
django-filter==2.4.0
django-recaptcha==2.0.6
django-thumbs==0.4
django-widget-tweaks==1.4.8
gunicorn==20.1.0
humanize==3.11.0
idna==3.2
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
